+++ package.json
@@ -21,5 +21,6 @@
   "dependencies": {
     "@clerk/nextjs": "^6.11.0",
     "zod": "^3.23.8"
-  }
+  },
+  "packageManager": "pnpm@9.15.0+sha512.76e2379760a4328ec4415815bcd6628dee727af3779aaa4c914e3944156c4299921a89f976381ee107d41f12cfa4b66681ca9c718f0668fa0831ed4c6d8ba56c"
 }
