+++ packages/server/package.json
@@ -17,7 +17,7 @@
   "dependencies": {
     "@clerk/express": "^1.3.41",
     "@prisma/client": "5.19.1",
-    "@trpc/server": "11.0.0-rc.724",
+    "@trpc/server": "11.0.0-rc.833",
     "bcryptjs": "^2.4.3",
     "cors": "^2.8.5",
     "dotenv": "^16.4.5",
